requirement.txt
absl-py==0.7.1
bleach==1.5.0
chardet==3.0.4
cython==0.29.8
grpcio==1.21.1
html5lib==0.9999999
idna==2.8
markdown==3.1.1
protobuf==3.7.1
pyyaml==5.1
requests==2.22.0
tqdm==4.46.1
tensorboard==1.6.0
tensorboard-pytorch==0.7.1
tensorboardx==1.1
urllib3==1.25.3
werkzeug==0.15.4
